Get all controller specs passing
Mostly we're dealing with actual spec errors that fell through the transpec process.  We also have to deal with some differences in how rails 3.0 protected attributes and the extracted gem work.  Seems that rails 3.0 would let us not specify attr_protected or attr_accessible on a model and we'd be able to use it "normally", but in rails 4.2 with protected_attributes all our models need to have a whitelist to allow any mass assignment.

I've not given much thought to how we apply this to the models as At Some Point Real Soon™ we'll be using the strong params approach.