Use dummy objects in spec/lib instead of doubles
The pain here is trying to stub methods that don't exist - a thing that RSpec 3.2 no longer allows you to do.  In search_results_setup_spec it's simple - we just implement the method on a fake object and use that instead of Object.extend(SearchResultSetup).

In search_spec it's more complex as we're trying to stub methods called in a block that is evaluated in a strange double context by solr.  We need to be able to stub the methods we know we care about that sunspot provides via its DSL (fulltext, facet, etc...) but still be able to call the methods that our block has access to in our Search object (options, etc...).  How sunspot does this is to establish a context for the block that includes the calling context, we try to do something similar, but more lightweight (e.g. not as robust).  It's horrible, but that's the price of DSLs and stubbing.