Don't enforce case-sensitivity for admin usernames
As of Rails 4.2 enforcing case-sensitivity for usernames is broken in
Authlogic due to the commit rails/rails@a38e957. The app change was for
security reasons since at the time Authlogic was using LIKE in the
query generation. However this was fixed in binarylogic/authlogic#339.

- binarylogic/authlogic#444
- binarylogic/authlogic#453

We should leave it as case-insensitive as this is better UX