Make Evergreen JS specs run again
When we require evergreen/rails we tell evergreen to mount itself as an engine via our standard rails app.  However, this means that it is forced to run via ssl.  This doesn't actually work when running the specs because the default dev server is not running ssl and we don't have the same extension points for forcing the app to run via ssl like we have with cucumber.  The solution is to roll back the changes to Evergreen that evergreen/rails does and just use the default Evergreen app without mounting it via the rails app.