Fixup moderation feature for Capybara 2.x
Capybara 2.0 and later is much more strict on matching text - it won't
match text in unselected <option> tags for example so we need to actually
select those options before using have_content.

Similarly, matching on the value attribute for options is not supported so
we need to use the label text for the empty value option.