Defer requiring attr_encrypted until Active Record has loaded
I think why attr_encrypted is breaking is that Bundler is requiring it
during the application booting phase before Active Record has loaded and
the guard clause in the attr_encrypted is preventing it from mixing in
correctly. The initializer failed to work because the file has already
been required so wasn't required a second time.