Predefine attribute methods for attr_encrypted
The attr_encrypted gem defines attribute accessors for attributes that
don't exist. Since at this point Active Record may not have defined the
methods it means that these methods are overriding the built-in methods.

To fix this we can predefine the methods to prevent attr_encrypted from
defining this attribute accessors.