Remove signatures_pre_encryption table
We won't need this anymore in the new version of the app, but we should take care when running this on a previous version of the app with data in it as we will lose that data if it's not already been emptied out.  The migration tweaks the definition slightly to give the indexes names that don't clash with the indexes on the real signatures table - I suspect this isn't really a problem on the production db though as we'd have noticed.