Rewrite steps for clarity
These steps were failing after changing to PostgreSQL because of fractional
timestamp support meaning that email_requested_at was later than the time
specified in the initial step.

Refactored to create two new steps that better explain the purpose of the
test rather than just checking to see if the record can be found.