Break creation of petitions into stages on the server
We extract a bunch of the code from petitions controller into a StagedPetitionCreator model class.  This wraps a petition to provide a staged interface.  We're only using new and create, but we work out what stage we are on based on the error state of the object.  To allow for pressing the "back" button to flip between stages we allow for passing an explicit "stage" param that takes precedence to the error state of the object.

In the new template we render the appropriate form UI for the stage using the partials we extracted in the previous commit.

Note that because this object wraps a petition object and we don't want to make the forms work on the underlying petition we delegate alot of methods from the StagedPetitionCreator to the Petition instance.  The one thing we don't delegate is the magic method that makes fields_for and nested_attributes_for work, so instead of getting ``creator_signature_attribtues[...]`` we now get ``creator_signature[...]`` and have to massage the params and forms to get things to work.  Notably:

1. we delete ``creator_signature`` params and re-store them as ``creator_signature_attributes`` params in ``StagedPetitionCreator#sanitize_params`` and update our specs to reflect this
2. we have to explicitly provide the object to our ``fields_for :creator_signature`` calls in the view as it can't be found automatically anymore: ``fields_for :creator_signature, petition.creator_signature``

This is a first pass, still more to come to make it functional.