Only create the creator_signature when we need it
If the creator_signature is always present then when we try to save the petition after submitting the "Petition details" UI we will fully validate it and the view is rendered with errors.  This isn't great as we end up telling the user about errors on the object before we've let them create it.

We do two things to avoid this:

1. only render the hidden UI if the petition has a creator_signature, this means the first petititon details form won't send any creator_signature attributes.
2. only instantiate the creator signature when we render form elements for providing it's details - using the creator_signature! method, instead of always building one in the build_petition method.

Coupled together this means we don't end up creating an instance to be validated until we need it.