Propagate stage and user movement through the params
We need to know what stage the user was on and what their choice was (to go to the next step or to go back) before we correctly work out what action to take and what UI to show them in response.  If the user clicks "back" we should always take them to the previous step, no matter the state of their current form; on the other hand if they click "next" we need to validate what they've done and only advance if those elements are correct.

The ``StagedPetitionCreator`` is now reasonably complex, and could probably do with being refactored somewhat - there may be other objects lurking inside it.  It might be worth separating out the stage calculation from the object presented to the form - we're mixing some concerns here.