Sort out tabindexes on staged petition create forms
It seems we want our forms to start at index 5 (because of existing tabindexed elements in the layout) so we increment(4) before rendering any of the partials.  As each partial calls increment() to get its first tabindex we'll end up with 5 as the first tabindex on the UI.

We also take the increment(0) calls on all the back/next buttons off and replace with increment() calls so that they have distinct tabindexes.