Override build/create methods for Sponsor signature association
This lets us default the petition and email on the signature which should be the same as the ones on the sponsor itself.  We have to cover build_signature, create_signature, and create_signature! as none of these are implemented in terms of themselves, but even if they were, there's no guarantee they always would be.  A small price to make sure our code always does what we expect.