Feature for emails sent to petition creators about sponsors
They should get an email when their sponsors sign, counting down to the threshold.  The final sponsor should not include the threshold, but should tell them their petition is in moderation.  After moderation they shouldn't get any more sponsor emails.

We add a step to allow marking all email for a given address as read.  It's useful to be able to do this because we want to be able to ignore all the emails a creator is sent before we then initiate another which will otherwise look almost the same.