Refactor search functionality
Refactors search functionality to get rid of custom sunspot search
code.

Department search is not 100% working (ie. search counts do not match
the results) and lack unit testing. That is intentional because we are
getting rid of the department functionality.

Preventing search with empty search term was also removed to
simplify the code. I saw no point preventing that, it is not an
expensive query and user can see all petitions anyway by clicking
a link in the site.