Add new "staged" models to represent what each stage needs
These wrap a petition and add validations for the stage that they represent.  So a petition in the "creator" stage will be valid if the creator details validations pass; the rest of the petition could be invalid, but we don't care about that at this stage.

These objects are somewhat complex as we've had to replicate some behaviour that our models have.  For example errors from creator_signature get propagated up to the petition by nested_attributes, active model doesn't provide this so we've had to implement it ourselves.