Remove Sunpot hacks and assume an instance is running
To aid in the movement towards using PostgreSQL full-text search as a
replacement for Solr remove all of the hacks that assume that we are
using Sunspot in our test environment and treat search as a black box.

We still need to run index updating an removal but these are gated with
respond_to? checks.