Rework staged petition creation to use "staged" models
The ``StagedPetitionCreator`` now works out what stage we have come from, what stage we're trying to get to, and what stage we're ultimately in based on validating the stage we've come from.  It uses the staged models to work out if we've passed or failed the stage and provides a new one we're advancing to the next stage.

This is still in-progress, there's a lot of complexity left in the ``StagedPetitionCreator`` that would probably be best extracted elsewhere.