Rework StagedPetitionCreator to use stage objects
These objects represent the state machine for how we go from stage to stage.  Each stage object knows two things:

1. what staged class to use to represent valid data for the stage it represents
2. how to transition from itself to the next (or back) stage

When asked to move it asks the staged class to validate if the stage is complete or not and uses that to decide if we should transition to the next (or back) stage or not.  (We almost always allow a "back" transition to succeed).

The StagedPetitionCreator is now more simple and just has to ask the correct questions of the stage objects.  The only complexity is if we managed to get to the 'done' stage but don't have a fully valid petition, in which case we find the earliest invalid stage and go back to that.