Remove user/deparment filtering from EmailReminder
It's possible that we should more fully refactor this now as we're always emailing every user about every set of petitions.  Perhaps we can now get away with only need to send one email and bcc-ing everyone?  Or maybe we need to be sending a different email to a different set of people.  This'll do for now, but we should discuss with @yahoopete.