Uniqueify sponsor_emails before turning into sponsors
If someone enters the same email address twice, we can easily strip it out so we might as well do this for them rather than complaining and making them do it.  Interestingly we were already doing this implicitly because during save the first duplicate would persist, but the second wouldn't (validates uniqueness of), but this somehow didn't fail the petition save.  Now we've made it explicit that we do this so there should be no surprises.