Clear old sponsors when building from sponsor_emails
This stops the sponsors array on the petition being built up with extra copies each time we validate.  We also drop our sponsors array on the Staged::Sponsors model in favour of delegating directly to the petition's sponsors association.  This has the added benefit of letting us use sponsors.build which will set the petition on the sponsor for us.