Move publishing logic into Petition
It's not clear the approach we're taking in this app as we have some business logic in the model and some in the controller, but this feels like it would be best in one place so we don't have to duplicate it when we want to publish a petition elsewhere.  We've got some duplication in the tests, but that's a small price.