Fix moderaion steps that relied on ``petition.department``
It's unclear why these didn't break earlier.  The step we changed was around checking that hidden petitions are no longer viewable and we were doing this by going through the departments page.  As our petitions no longer neccessarily have departments this failed when trying to visit the page.  Instead we go though the search page, which means we have to change the ``I should see the petition "<title>"`` step to check for a link with the title, not just the text being on the page.