submit is not a good name for a stage, use replay-petition
This better expresses what the stage does; we replay the petition details that the user has provided and allow them to go back to change them if necessary.  Shortly we will be dropping the t&cs from the UI in this stage and moving it to live before the creator step so looking ahead 'submit' will become an even worse name for it so renaming it now makes sense.