Add a new stage: ReplayEmail
This stage validates the email of the creator signature.  The intention here is to replay the users' email instead of asking for email confirmation earlier.  If the email is wrong we let the user change it.

This is now the final step before we get try to persist the petition to the DB.

There is a wrinkle here.  The 'creator' stage and the 'replay-email' stage both validate the email of the creator_signature so it's impossible (currently) from the 'done' stage to go back to the correct stage because both 'creator' and 'replay-email' would be a candidate.  We just pick the earliest one which will be 'creator'.  It would be good to work out a solution here.

However, in practice we'd never submit a POST request with stage='done' and move='next' so this situation will most likely never occur.  If the user is on 'replay-email' and breaks the email, the ReplayEmail stage will fail validation so we won't go to 'done' and we won't rewind to 'creator'.  If the user somehow crafts a POST with stage 'done' then we don't know where they came from so throwing them back to 'creator' doesn't seem so unreasonable.

An alternate approach might be the best thing here.