Add the UI for the replay email stage
We have a reasonable abstraction within the controller / staged petition creator for working out the controller side, but what view elements we render out are becoming increasingly complex.  Particularly as both the creator and replay-email stages will render UI for creator_signature[email] so we don't want to render the hidden version.  Might need to revisit the hidden elements logic.