Add Staged::PetitionSigner
This gives us staged creation of signatures with a "signer" stage and a "replay-email" stage.  Currently not hooked into the controller / UI.  We will want to revisit the implementation  as it's very similar to Staged::PetitionCreator, but it's not clear that inheritance is the best solution, particularly for the internal Stages module.  Will ponder.