Extract validation of multiple signers to staged signatures
We need to perform the multiple signers validation on the staged models so that the user doesn't get a nasty surprise after going through the previous two stages and we tell them "oh, no, you've already signed".  Also because these validations weren't on the staged models we wouldn't know which stage to go back to, nor would we put the error messages on screen; so this kills two birds with one stone.