Move ``Staged::Petition`` into ``Staged::PetitionCreator``
These objects were only ever used by the ``Staged::PetitionCreator`` so it makes sense that they reflect that more strongly by living in that namespace.

The problem is that because ``Staged::Petition`` was a module and ``Staged::PetitionCreator`` was a class rails autoloading started to play up and references in ``Staged::PetitionCreator::Stages`` to ``Staged::PetitionCreator::Petition`` were resolving to ``Petition`` (the top level AR model, not our AM model for staged validation).  It's unclear why the autoloading was breaking, but as it was fine with a module but not a class we also shuffle things around so that ``Staged::PetitionCreator`` is a module.  The object that used to have that name is now a ``Staged::PetitionCreator::Manager`` that we get via ``Staged::PetitionCreator.manage`` instead of ``Staged::PetitionCreator.new``.

This has the nice side-effect of making the instance variable name for the instance of this object we create in the petitions controller (``@stage_manager``) feel more appropriate.  Slight issue is that we have ``Staged::StageManager`` which our ``@stage_manager`` is not an instance of, but it does use one internally and it performs that role of managing the stages for the controller so I think we'll be fine.