Avoid short-circuiting ``valid?`` in Staged models
In the classes where we have overridden ``valid?`` to also validate a child object (the ``HasCreatorSignature`` mixin and ``Sponsors`` model) we were doing:

    def valid?
      super && some_other_method_that_calls_valid_on_a_child_object
    end

The problem here is that if ``super`` returned false we short-circuit the expression and don't run the other method and so don't validate the child object.  Instead we use ``validate`` to call the other method as part of the standard validation process.

In practice it wasn't a problem as I neither of the affected classes had any validations on the "parent" so ``super`` would never fail.  However it's not a good pattern to leave lying around.