Describe new sponsor validation flow in cucumber
Sponsors now have to verify their email address and we want to use the "ask for email, then replay" flow we use for normal signatures.  The email we send them to verify will also include the exact text of the petition, just in case it's different to what they were told by the creator.