Extract signature verify logic from controller to model
We add Signature#validate_from_token! to house this: it compares the token and if it's the same we go to validated state, reset the token to nil, timestamp the model and return true.  Otherwise return false.

We've also added a guard clause that wasn't present in the controller logic where we return false if the signature was already validated, or if it has no perishable token (given the flow of this method these two things should be interchangable, but you never know).