Force by_role scope to be ordered by id
When no ORDER BY clause is specified in SQL then there is no guarantee
of the order of the returned records. This was causing intermittent
failures in the email_reminder_spec.rb file because of the order of the
arguments passed to the AdminMailer.threshold_email_reminder method

This fixes those intermittent failures by forcing the by_role scope to
always return users ordered by id.