Make PerishableTokenGenerator more general
Instead of just adding the before_create callback and method for a field called perishable_token it provides a "macro" ``has_perishable_token`` that you should call to generate that callback and method.  ``has_perishable_token`` takes an argument (``called``) that we can use to change the name of the token we want to generate (it defaults to ``perishable_token``).  This allows us to re-use the generator to do both ``perishable_token`` and ``unsubscribe_token`` on ``Signature``.