Rework signature and sponsors controllers to use new token
The token now belongs to the petition so it only identifies the petition not an individual sponsor.  The flow in sponsors controller becomes significantly simpler as we can't actually detect the state of the sponsor in thank-you / new / sponsored actions anymore, because we don't have a sponsor object to look at any more.