Add archived petitions controller and views
The views are based on the current views for petitions and are likely to
be changed when the final design is implemented so not too much effort
went into their implementation.

The petition controller will redirect to the archived petitions controller
if a petition with the id doesn't exist but an archived petition does.