Force Sunspot to commit changes to Solr
Since the index updates can be time consuming, Sunspot waits until the
end of the request to send those updates. This means we need to force
the index updates to be committed so that the tests will pass with a
fresh copy of the database as it is when running on the CI server.