Add cookie information message
When a user visits a site for the first time show them a message about
cookies and link to the cookies section on the help page. On any
subsequent page views the cookie message is hidden by the presence of
a cookie called `seen_cookie_message` being set to `yes`.

This follows the pattern establised on GOV.UK and in the govuk_template
gem[1] but the implementation is slightly different in that we're not
using JavaScript to determine whether to show the message or not and
we're also setting the expiry date on the cookie that stores whether the
user has seen the message or not to 1 year and not the standard 28 days.

We may need to revisit this when we look at performance and caching.

[1]: https://github.com/alphagov/govuk_template
[2]: https://www.pivotaltracker.com/story/show/95857040