New feature to describe local petition functionality
We can search from the homepage with a postcode to see a list of petitions that have signatures from within our constituency.

This feature and new step definitions do point us in a specific implementation direction (storing constituency_id on the signature) but we may need to revisit as we go.