Add Signature#set_constituency_id
This calls the constituency method to fetch the constituency and set its id onto constituency_id on the instance.  We were going to set this as a ``before_create`` lifecycle event, but it caused way to many specs to break and the possible solutions didn't feel nice (global stub_request for the default postcode in the factory, or explicitly stubbing in every spec where we create a signature).  For now it's something we'll have to call explicitly.