Increment constituency signature count after validating signature
When we call Signature#validate! it will in turn call ConstituencyPetitionJournal.record_new_signature_for to increment the count for that signature's petition and constituency.  We also refactor the validate! method slightly (thanks to a discussion with @pixeltrix) to minimise the SQL statements used.

While we don't make sure that all validated signatures in our specs or features are created as pending and moved across to validated via ``#validate!`` we do update the ``constituency_search_steps.rb`` to do it that way.  Currently they're the only tests that might care.