Added Petition.popular_in_constituency
This takes a constituency_id and an optional number for how many to return.  It does the query to fetch the set of petitions that have signatures from that constituency ordered such that those with the most signatures appear first.  The constituency_signature_count is added to the petition object by this query.

Note that this method uses scopes internally, but returns a flat array of results, not a scope for further chaining.  The query is complex so further chaining of methods might break it, so this makes it a bit safer to use.