Use locale keys to correctly pluralize signature counts
The i18n gem has a feature where if you pass it an option called count
then it will use the sub key called one if the value of count is 1 and
if the value of count is more than 1 then it uses other. Coupled with
the use of the html key to mark the output as HTML-safe allows us to
localize the language and formatting if needed at some point.