Make signature counts live
Rather than re-counting the signature counts every 5 minutes update them
after a signature is validated. By using a query of this form:

  UPDATE petitions SET signature_count = signature_count + 1 WHERE id = ?

we can avoid race conditions because the update is atomic.

Unfortunately the factories were pretty reliant on the scheduled task to
correct the signature counts when creating the records - hopefully I've
managed to catch all the places where this was happening.

https://www.pivotaltracker.com/story/show/96434874