Remove chronic usage in specs
Although the explicit requirement of the chronic gem was removed
in 05f360a it's implicitly required by the whenever gem so some
specs were still using it. This commit changes those specs to not
rely on chronic so we can look at other ways of scheduling tasks
and not have the specs fail if we remove the whenever gem.