Merge AppConfig and SystemSetting into Site model
Rather than have two places to configure things it makes sense to combine
these into a single place. This Site model can also serve some additional
purposes such as holding timestamp for globally clearing the cache as well
as protecting the website behind HTTP basic authentication and disabling
the website so that it will return a 503 Service Unavailable response.

Since we only have at most a handful of settings it makes more sense to
store them in explicit columns than cater for an infinite set.

    class ApplicationController < ActionController::Base
      before_filter :reload_site

      protected

      def reload_site
        Site.reload
      end
    end

This will also be useful for when we implement 'Russian Doll' caching - by
incorporating the updated_at column in the cache key we can create a global
reset that we can both hook up to an admin function or call programmatically
in a deploy script if we need to blow away any existing cache.

Some notes on implementation:

1.  The model is cached in a thread local for performance and then reloaded
    in the application controller to allow for admin updates.

2.  The before_remove_const method is a hook that Rails calls when unloading
    classes before reloading them in development mode. This allows us to
    clear the cached instance in the thread local, otherwise we would end up
    with an instance that has a stale class definition.

3.  By using first_or_create to load the instance we can automatically seed
    the database with the correct values by defining the appropriate defaults
    on the respective columns. Changing to this exposed some failures in our
    cucumber test setup - it was defaulting to a threshold_for_debate level
    of 0 so I had to correct for that.

5.  I've added explicit settings for the debate and response threshold to
    make clear what each one is for.