Move Faker gem require inside the task block
When we run `rake db:migrate` in production it loads all of the tasks
but since `gem 'faker'` is only in the development and test bundler
groups it blows up with a LoadError. We can fix this by moving the
require inside the body of the task because Ruby's require will detect
that the gem has already been loaded.