Rework how we show debate outcomes and responses on petition show
They need to appear in the "10,000 signatures" / "100,000 signatures" info blocks instead of the "will get a response" / "will be debated" info if the relevant stuff is present.  The order of this block needs to be different depending on the state: 100,000 comes first if debate outcome is present, 10,000 otherwise.

We've moved all this logic into the ``shared/about_petitions`` partial (which we rename to ``petitions/threshold_details`` as that's more representative) which is shared by the "you've signed" page at the end of the signing flow.  It's not clear if this page *should* inherit this behaviour or not, if not we'll just split it off and have two partials.