Tidy up spec definitions
It's preferable to use the singleton describe method on the RSpec
constant and explicit example type metadata. This allows us to turn
off some features that will either be off by default in RSpec 4 or
otherwise not available.

The profiling was disabled because it doesn't really show anything
of use for our specs and was just noise.