Use explict exception classes with raise_error
It's good practise not to use the raise_error matcher without a specific
exception class because it can mask errors. One such case of masking was
uncovered in the signatures_controller_spec.rb where not passing the
petition_id to the params actually caused a url generation error when
constructing the mock request and never actually hit the controller.