Tidy up helper usage in mailers
The ActionView::Helpers::NumberHelper is focussed on HTML output but we're
using it for subjects which are just strings. Rails 4.1+ extracted the
number formatting helpers to Active Support so we can just include that
module directly in our mailer.

Also the `add_template_helper` method is a private Rails API method - the
correct method to use is `helper`.