Add ``be_usec_precise_with`` matcher for rspec
This should help with comparing time objects that do a round trip with the DB.  The DB stores seconds to 6 digits (microsecond precision), but depending on the OS our time objects could have more precision digits.  The helper uses ``be_within(0.000001.seconds).of`` to make sure that we only compare the two times at microsecond precision (and cater for the fact that the DB truncates extra digits rather than roungs them).