Rename Petition#description => Petition#additional_details
We'd done it as "Supporting details" in the view and wanted to propagate this change to the db and model too.  We decided that the word "supporting" might make people think this was text to be sent to the sponsors or signatories (we also talk about them as "supporters") so we decided to change it to "Additional details" instead across the board.