Track deployment progress and notify New Relic when complete
So we can track whether a deploy has negatively (or positively) affected
the performance of the website record them in New Relic. We can't do this
the traditional way by running the CLI command or cap task since we're
running across any number of servers so we poll the AWS API to get a
progress of the deployment and notify New Relic ourselves when the deploy
has successfully completed.