Rename ``Petition#action`` to ``Petition#background``
We'd done it in the view so it makes sense to match up in the db and model too.

Instead of renaming the index_petitions_on_action index, we drop it and create the index_petitions_on_background because the original index was created against the wrong field when originally added.