Provide base class ``EmailPetitionSignatories::Job``
This is where ``run_later_tonight`` lives as it didn't make sense on ``EmailPetitionSignatories`` which wasn't actually a job.  We make `EmailDebateOutcomesJob`` and ``EmailThresholdResponseJob`` subclasses of this new base class which allows us to move the ``worker`` method out too as it was the same implementation in both subclasses.

The worker feels a little redundant now that we have a base class where everything it does could live.  However I'm inclined to keep it because it lets us test a simple object rather than having to play with the slightly odd ``ActiveJob::Base`` api for creating objects (you have to do it via the queue as there's no initialize).