Rework moderation features and steps
The major change is that we won't have admin/petitions/edit be the UI for moderating - it'll just appear on the show page.  We also don't toggle the action by the button the user pressed, but by making them choose from two radio buttons and then press a single "Go" button.