Add the UI for moderation to ``/admin/petitions/<id>``
Because we don't use an ``edit`` action on moderations controller but instead just render the ``admin/petitions/show`` when the approve/reject doesn't work we need to be explicit with folder names when rendering partials in this view.  If we don't then they start looking for things in app/admin/moderation or app/admin/petitions at the wrong point.