Remove response_required flag and rename internal_response
We remove public_response_required as it's not useful anymore.  Moderators will know if a public response is required by the number of signatures and the various queues we'll put the petitions into.  There is no need for them to flag a petition to get someone to add a response.

We rename internal_response to admin_notes because that better reflects what is stored in there (at some point this might be a has_many association to allow an audit trail).