Move the 'change rejection status' UI out of `edit_response`
The take down UI and the change rejection status UI are basically the same and target the same backend action.  It's easy enough to toggle the view in take_down/show based on if the petition is already ``.rejected?`` so let's just do that rather than having a separate action just for this UI.

We also remove the status guard on the take_down controller - the old action didn't have a guard so there's no reason for this new one to have it.  Particularly as we need it to work with rejected petitions too now.