Remove 'email_signees'
We no longer have to let the admins choose whether to email signees or not when editing the government response.  In the past they could do other things via this UI and backend, so sometimes they absolutely didn't want to email the signees because they hadn't changed anything to do with the government response.  This UI and backend now only deal with the government response so there's no need to give a choice; they'll always be intending to email the signees.