Change destination after submitting government response
According to the design we should go back to the form and show a flash message.  It's not clear if we actually expect the normal show page to embed this form (probably defaulting to opening the correct form based on the state of the petition - kind of like we do for the moderation form).