Standardise button text and flash message for admin
For admin actions that will email all the signatures on a petition make sure we use 'Email <n> signatures' on the submit button and (where it's true) set a flash message of 'Email will be sent overnight' to let them know it's not instant.  This is: debate_outcomes, government_response, and schedule_debate_date.  Other actions do send email, but they don't send email to everyone (take_down and reject for example only email creator+sponsors) or they don't email overnight.

We move the generation of the "Email <n> signatures" submit button into a helper, this way we are able to make sure we also always have a data-confirm popup for these actions.  Sending an email to that many people should be something we do with a little care and not via a slip of the mouse.