Make controls on admin all petitions autosubmit filter form
When you choose a new state to filter the all petitions list by, or a new number of petitions to see per page, we want the user to not have to press go.  We do this with a simple jquery enhancement that attaches to any select element with a data-autosubmit attribute and adds a onchange handler to submit the form the select lives in.

We leave the 'Go' button just in case (but perhaps we should hide it?).

Notes of interest:

1. We test this only via cucumber but this means we had to change some of our inspection steps to use capybara matchers instead of rspec matchers.  Without using the capybara matcher we don't neccessarily wait for the page refresh so might assert on the previous page contents.

2. We have an old version of jquery so we need to do `$(this).closest('form').submit();` - a newer jquery would let us do `$(this).form().submit();`.