Update debated_petition factory to set the petition
The factory was previously creating 2 petitions at once as it has debate_outcome as an association which in turn has a petition association.
One of the feature tests seems to have counted on these extra petitions being created, so have fixed it.
Have also set the default date to be 1 day ago, so that it would pass validation.